Wait to start runtime until 'configurationDone'
The mock debug engine currently starts the runtime as soon as a `launch`
request is received.  This can cause problems such as missed breakpoints
if configuration is not yet complete.  This commit changes the logic so
that the runtime will not actually begin execution until the
`configurationDone` message is received from the host.